+++ client/src/screens/hub/HubMainScreen/components/ContextBox/ContextBox.js
@@ -310,20 +310,26 @@ class ContextBox extends Component {
         const colorObj = Color(color);
 
         let active = false;
-        if ((
-          focusedCircle.runHash === run.run_hash
-          && focusedCircle.metricName === metric?.name
-          && focusedCircle.traceContext === contextHash)
+
+        if ((this.context.runs?.meta?.params_selected 
+          && ((focusedCircle.runHash === run.run_hash) || (focusedMetric.runHash === run.run_hash)))
           || (
-            focusedMetric.runHash === run.run_hash
-            && focusedMetric.metricName === metric?.name
-            && focusedMetric.traceContext === contextHash)) {
+            focusedCircle.runHash === run.run_hash
+            && focusedCircle.metricName === metric?.name
+            && focusedCircle.traceContext === contextHash)
+            || (
+              focusedMetric.runHash === run.run_hash
+              && focusedMetric.metricName === metric?.name
+              && focusedMetric.traceContext === contextHash)) {
           active = true;
         }
 
-        if (focusedCircle.runHash === run.run_hash
-          && focusedCircle.metricName === metric?.name
-          && focusedCircle.traceContext === contextHash) {
+        if ((this.context.runs?.meta?.params_selected 
+          && (focusedCircle.runHash === run.run_hash))
+          || (
+            focusedCircle.runHash === run.run_hash
+            && focusedCircle.metricName === metric?.name
+            && focusedCircle.traceContext === contextHash)) {
           expanded[JSON.stringify(traceModel.config)] = true;
         }