+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+- Add circles and interactions on parallel coordinates plot (roubkar)
 - Add parallel coordinates plot chart number and config as title (roubkar)
 - Recover grouping popup focus in controls sidebar after button click (roubkar)
 - Make context table row heights lower (gorarakelyan)